Add confirm dialog when deleting cards
Summary: Ask for confirmation

Test Plan: None

Reviewers: warren

Reviewed By: warren

Subscribers: a.h, qromanek, trickywolffe, unzunz

Projects: #edge-yl

Differential Revision: http://phab.calblueprint.org/D171

+++ app/assets/javascripts/components/groups/groups_card.es6.jsx
@@ -19,7 +19,10 @@ class GroupsCard extends Component {
   // Helpers
   // --------------------------------------------------
   deleteGroupFromConference() {
-    ConferenceActions.deleteGroup(this.props.group.id);
+    var response = confirm('This action cannot be undone.');
+    if (response) {
+      ConferenceActions.deleteGroup(this.props.group.id);
+    }
   }
 
   // --------------------------------------------------
